Addressed code review comments.
